refactor article-page.js and related/index.js - add src/components/related/list.js
